+++ R/admixcor.R
@@ -6,6 +6,7 @@ admixcor <- function(
                      delta = 0.01,
                      Q_type = c('kmeans', 'random', 'uniform'),
                      L_type = c('identity', 'uniform', 'diagrandom', 'random'),
+                     L_algorithm = c('original', 'nnls', 'bvls', 'glmnet'),
                      tol = sqrt( .Machine$double.eps ), # 1e-15
                      nstep_max = 100000,
                      report_freq = 1000
@@ -13,7 +14,8 @@ admixcor <- function(
     # process options
     Q_type <- match.arg( Q_type )
     L_type <- match.arg( L_type )
-
+    L_algorithm <- match.arg( L_algorithm )
+    
     # number of individuals is dimensions of Theta
     n <- nrow( Theta )
 
@@ -57,7 +59,7 @@ admixcor <- function(
         
         # apply the updates, one at the time
         R1 <- update_R( ThetaSR, Q0, L0 )
-        L1 <- update_L( ThetaSR, Q0, R1, gamma, I )
+        L1 <- update_L( ThetaSR, Q0, R1, gamma, I, algorithm = L_algorithm )
         Q1 <- update_Q( ThetaSR, L1, R1, delta, I )
         
         # calculate step sizes, to assess convergence